Updated FDR cutoff and ensembl_to_id functions
Previously, the FDR filtering function was limited to <.05. It has now been renamed from f05 to fdr_cutoff and the cutoff value can be user adjusted, but it does default to .05.

The ensembl_to_id function was found to contain errors using cbind. It has now been changed to use the merge function to combine the data frames. The amount of information provided has also been decreased from ensembl_transcript_id, external_gene_name, ensembl_gene_id, and description to ensembl_transcript_id_version, external_gene_name, description.

It might be worthwhile in future updates to default to the current attributes, but allow for user editing to what is retrieved.