Updated ensemble_to_id and sleuth_object_result.
Previously, the sleuth_object_result function would combine all ID's passing the fdr_cutoff function into one data frame. This has now been changed and IDs passing this threshold are now assigned an independent variable at the global scope.

Three new parameters have been introduced: all_data, sig_data, and q_max. The all_data parameter defaults to TRUE and permits the creation and assignment of all WT results. The sig_data is similar to the all_data variable, but is specific to only the IDs that pass the fdr_cutoff. The q_max value is passed to the fdr_cutoff function and defaults to .05.

The ensemble_to_id function was renamed to ensembl_to_id, which is in line with the actual ensemble name. The others changes were printing related that are evoked when users do not have a mart object named `mart` in the global enviroment.