Updated ensembl_to_id and added volc_plot function
Modified the ensembl_to_id function to properly arrange the columns when data frames of different sizes and used. This was done as an error arose when processing lrt results.

Added in the volc_plot function to rapidly create a colvano plot based on sleuth results. Although a function for this already existed, it did not contain the cutoffs or different colours that this function introduces.